refactoring tracks with superclasses for name and pattern follow
