+++ src/apps/sequencer/model/NoteTrack.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "BaseTrackPatternFollow.h"
 #include "Config.h"
 #include "Types.h"
 #include "NoteSequence.h"
@@ -7,14 +8,14 @@
 #include "Routing.h"
 #include "FileDefs.h"
 #include "core/utils/StringUtils.h"
+#include "BaseTrack.h"
 
 
-class NoteTrack {
+class NoteTrack : public BaseTrack, public BaseTrackPatternFollow {
 public:
     //----------------------------------------
     // Types
     //----------------------------------------
-    static constexpr size_t NameLength = FileHeader::NameLength;
 
     typedef std::array<NoteSequence, CONFIG_PATTERN_COUNT + CONFIG_SNAPSHOT_COUNT> NoteSequenceArray;
 
@@ -59,12 +60,6 @@ public:
     //----------------------------------------
     // Properties
     //----------------------------------------
-
-    // trackName
-    const char *name() const { return _name; }
-    void setName(const char *name) {
-        StringUtils::copy(_name, name, sizeof(_name));
-    }
 
     // playMode
 
@@ -270,42 +265,6 @@ public:
         str("%+.1f%%", noteProbabilityBias() * 12.5f);
     }
 
-    // patternFollow
-    Types::PatternFollow patternFollow() const { return _patternFollow; }
-    void setPatternFollow(const Types::PatternFollow patternFollow) {
-        _patternFollow = ModelUtils::clampedEnum(patternFollow);
-    }
-
-    void setPatternFollow(bool trackDisplay, bool trackLP) {
-
-        if (trackDisplay && trackLP) {
-            setPatternFollow(Types::PatternFollow::DispAndLP);
-            return;
-        }
-
-        else if (trackDisplay) {
-            setPatternFollow(Types::PatternFollow::Display);
-            return;
-        }
-
-        else if (trackLP) {
-            setPatternFollow(Types::PatternFollow::LaunchPad);
-            return;
-        }
-
-        setPatternFollow(Types::PatternFollow::Off);
-
-        return;
-    }
-
-    void editPatternFollow(int value, bool shift) {
-        setPatternFollow(ModelUtils::adjustedEnum(patternFollow(), value));
-    }
-
-    void printPatternFollow(StringBuilder &str) const {
-        str(Types::patternFollowName(patternFollow()));
-    }
-
     // sequences
 
     const NoteSequenceArray &sequences() const { return _sequences; }
@@ -342,7 +301,6 @@ private:
     }
 
     int8_t _trackIndex = -1;
-    char _name[NameLength + 1];
     Types::PlayMode _playMode;
     FillMode _fillMode;
     bool _fillMuted;
@@ -355,7 +313,6 @@ private:
     Routable<int8_t> _retriggerProbabilityBias;
     Routable<int8_t> _lengthBias;
     Routable<int8_t> _noteProbabilityBias;
-    Types::PatternFollow _patternFollow;
 
     NoteSequenceArray _sequences;
